@@ -1,21 +1,27 @@
 name: android_intent
 description: Flutter plugin for launching Android Intents. Not supported on iOS.
-author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/flutter/plugins/tree/master/packages/android_intent
-version: 0.3.2
+version: 2.0.0-nullsafety.2
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.androidintent
-    iosPrefix: FLT
-    pluginClass: AndroidIntentPlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.androidintent
+        pluginClass: AndroidIntentPlugin
 
 dependencies:
   flutter:
     sdk: flutter
-  platform: ^2.0.0
-  meta: ^1.0.5
+  platform: ^3.0.0-nullsafety.4
+  meta: ^1.3.0-nullsafety.6
+dev_dependencies:
+  test: ^1.16.0-nullsafety.13
+  mockito: ^4.1.3
+  flutter_test:
+    sdk: flutter
+  pedantic: ^1.10.0-nullsafety.1
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.2.0 <2.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5"
