@@ -1,15 +1,17 @@
 name: quick_actions
 description: Flutter plugin for creating shortcuts on home screen, also known as
   Quick Actions on iOS and App Shortcuts on Android.
-author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/flutter/plugins/tree/master/packages/quick_actions
-version: 0.3.2+2
+version: 0.4.0+12
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.quickactions
-    iosPrefix: FLT
-    pluginClass: QuickActionsPlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.quickactions
+        pluginClass: QuickActionsPlugin
+      ios:
+        pluginClass: FLTQuickActionsPlugin
 
 dependencies:
   flutter:
@@ -21,7 +23,10 @@ dev_dependencies:
   mockito: ^3.0.0
   flutter_test:
     sdk: flutter
+  integration_test:
+    path: ../integration_test
+  pedantic: ^1.8.0
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.5.0 <2.0.0"
+  sdk: ">=2.1.0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5"
