@@ -1,27 +1,33 @@
 name: share
 description: Flutter plugin for sharing content via the platform share UI, using
   the ACTION_SEND intent on Android and UIActivityViewController on iOS.
-author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/flutter/plugins/tree/master/packages/share
-version: 0.6.2+1
+version: 2.0.0-nullsafety.2
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.share
-    iosPrefix: FLT
-    pluginClass: SharePlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.share
+        pluginClass: SharePlugin
+      ios:
+        pluginClass: FLTSharePlugin
 
 dependencies:
-  meta: ^1.0.5
+  meta: ^1.3.0-nullsafety.6
+  mime: ^1.0.0-nullsafety.0
   flutter:
     sdk: flutter
 
 dev_dependencies:
-  test: ^1.3.0
-  mockito: ^3.0.0
+  test: ^1.16.0-nullsafety.13
+  mockito: ^4.1.3
   flutter_test:
     sdk: flutter
+  integration_test:
+    path: ../integration_test
+  pedantic: ^1.10.0-nullsafety.3
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.6.0 <2.0.0"
+  flutter: ">=1.12.13+hotfix.5"
+  sdk: ">=2.12.0-0 <3.0.0"
