@@ -1,6 +1,9 @@
 // Copyright 2017 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
+
+// ignore_for_file: public_member_api_docs
+
 import 'dart:async';
 
 import 'package:flutter/material.dart';
@@ -18,15 +21,28 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   final LocalAuthentication auth = LocalAuthentication();
-  bool _canCheckBiometrics;
-  List<BiometricType> _availableBiometrics;
+  _SupportState _supportState = _SupportState.unknown;
+  bool? _canCheckBiometrics;
+  List<BiometricType>? _availableBiometrics;
   String _authorized = 'Not Authorized';
+  bool _isAuthenticating = false;
+
+  @override
+  void initState() {
+    super.initState();
+    auth.isDeviceSupported().then(
+          (isSupported) => setState(() => _supportState = isSupported
+              ? _SupportState.supported
+              : _SupportState.unsupported),
+        );
+  }
 
   Future<void> _checkBiometrics() async {
-    bool canCheckBiometrics;
+    late bool canCheckBiometrics;
     try {
       canCheckBiometrics = await auth.canCheckBiometrics;
     } on PlatformException catch (e) {
+      canCheckBiometrics = false;
       print(e);
     }
     if (!mounted) return;
@@ -37,10 +53,11 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<void> _getAvailableBiometrics() async {
-    List<BiometricType> availableBiometrics;
+    late List<BiometricType> availableBiometrics;
     try {
       availableBiometrics = await auth.getAvailableBiometrics();
     } on PlatformException catch (e) {
+      availableBiometrics = <BiometricType>[];
       print(e);
     }
     if (!mounted) return;
@@ -53,48 +70,151 @@ class _MyAppState extends State<MyApp> {
   Future<void> _authenticate() async {
     bool authenticated = false;
     try {
-      authenticated = await auth.authenticateWithBiometrics(
-          localizedReason: 'Scan your fingerprint to authenticate',
+      setState(() {
+        _isAuthenticating = true;
+        _authorized = 'Authenticating';
+      });
+      authenticated = await auth.authenticate(
+          localizedReason: 'Let OS determine authentication method',
           useErrorDialogs: true,
           stickyAuth: true);
+      setState(() {
+        _isAuthenticating = false;
+        _authorized = 'Authenticating';
+      });
     } on PlatformException catch (e) {
       print(e);
+      setState(() {
+        _isAuthenticating = false;
+        _authorized = "Error - ${e.message}";
+      });
+      return;
     }
     if (!mounted) return;
 
+    setState(
+        () => _authorized = authenticated ? 'Authorized' : 'Not Authorized');
+  }
+
+  Future<void> _authenticateWithBiometrics() async {
+    bool authenticated = false;
+    try {
+      setState(() {
+        _isAuthenticating = true;
+        _authorized = 'Authenticating';
+      });
+      authenticated = await auth.authenticate(
+          localizedReason:
+              'Scan your fingerprint (or face or whatever) to authenticate',
+          useErrorDialogs: true,
+          stickyAuth: true,
+          biometricOnly: true);
+      setState(() {
+        _isAuthenticating = false;
+        _authorized = 'Authenticating';
+      });
+    } on PlatformException catch (e) {
+      print(e);
+      setState(() {
+        _isAuthenticating = false;
+        _authorized = "Error - ${e.message}";
+      });
+      return;
+    }
+    if (!mounted) return;
+
+    final String message = authenticated ? 'Authorized' : 'Not Authorized';
     setState(() {
-      _authorized = authenticated ? 'Authorized' : 'Not Authorized';
+      _authorized = message;
     });
+  }
+
+  void _cancelAuthentication() async {
+    await auth.stopAuthentication();
+    setState(() => _isAuthenticating = false);
   }
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-        home: Scaffold(
-      appBar: AppBar(
-        title: const Text('Plugin example app'),
-      ),
-      body: ConstrainedBox(
-          constraints: const BoxConstraints.expand(),
-          child: Column(
-              mainAxisAlignment: MainAxisAlignment.spaceAround,
-              children: <Widget>[
+      home: Scaffold(
+        appBar: AppBar(
+          title: const Text('Plugin example app'),
+        ),
+        body: ListView(
+          padding: const EdgeInsets.only(top: 30),
+          children: [
+            Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                if (_supportState == _SupportState.unknown)
+                  CircularProgressIndicator()
+                else if (_supportState == _SupportState.supported)
+                  Text("This device is supported")
+                else
+                  Text("This device is not supported"),
+                Divider(height: 100),
                 Text('Can check biometrics: $_canCheckBiometrics\n'),
-                RaisedButton(
+                ElevatedButton(
                   child: const Text('Check biometrics'),
                   onPressed: _checkBiometrics,
                 ),
+                Divider(height: 100),
                 Text('Available biometrics: $_availableBiometrics\n'),
-                RaisedButton(
+                ElevatedButton(
                   child: const Text('Get available biometrics'),
                   onPressed: _getAvailableBiometrics,
                 ),
+                Divider(height: 100),
                 Text('Current State: $_authorized\n'),
-                RaisedButton(
-                  child: const Text('Authenticate'),
-                  onPressed: _authenticate,
-                )
-              ])),
-    ));
+                (_isAuthenticating)
+                    ? ElevatedButton(
+                        onPressed: _cancelAuthentication,
+                        child: Row(
+                          mainAxisSize: MainAxisSize.min,
+                          children: [
+                            Text("Cancel Authentication"),
+                            Icon(Icons.cancel),
+                          ],
+                        ),
+                      )
+                    : Column(
+                        children: [
+                          ElevatedButton(
+                            child: Row(
+                              mainAxisSize: MainAxisSize.min,
+                              children: [
+                                Text('Authenticate'),
+                                Icon(Icons.perm_device_information),
+                              ],
+                            ),
+                            onPressed: _authenticate,
+                          ),
+                          ElevatedButton(
+                            child: Row(
+                              mainAxisSize: MainAxisSize.min,
+                              children: [
+                                Text(_isAuthenticating
+                                    ? 'Cancel'
+                                    : 'Authenticate: biometrics only'),
+                                Icon(Icons.fingerprint),
+                              ],
+                            ),
+                            onPressed: _authenticateWithBiometrics,
+                          ),
+                        ],
+                      ),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
   }
+}
+
+enum _SupportState {
+  unknown,
+  supported,
+  unsupported,
 }
