@@ -13,17 +13,16 @@ const String _backgroundName =
     'plugins.flutter.io/android_alarm_manager_background';
 
 // This is the entrypoint for the background isolate. Since we can only enter
-// an isolate once, we setup a MethodChannel to listen for method invokations
+// an isolate once, we setup a MethodChannel to listen for method invocations
 // from the native portion of the plugin. This allows for the plugin to perform
 // any necessary processing in Dart (e.g., populating a custom object) before
 // invoking the provided callback.
 void _alarmManagerCallbackDispatcher() {
-  const MethodChannel _channel =
-      MethodChannel(_backgroundName, JSONMethodCodec());
-
-  // Setup Flutter state needed for MethodChannels.
+  // Initialize state necessary for MethodChannels.
   WidgetsFlutterBinding.ensureInitialized();
 
+  const MethodChannel _channel =
+      MethodChannel(_backgroundName, JSONMethodCodec());
   // This is where the magic happens and we handle background events from the
   // native portion of the plugin.
   _channel.setMethodCallHandler((MethodCall call) async {
@@ -39,8 +38,10 @@ void _alarmManagerCallbackDispatcher() {
       exit(-1);
     }
 
+    // ignore: inference_failure_on_function_return_type
     if (closure is Function()) {
       closure();
+      // ignore: inference_failure_on_function_return_type
     } else if (closure is Function(int)) {
       final int id = args[1];
       closure(id);
@@ -52,14 +53,34 @@ void _alarmManagerCallbackDispatcher() {
   _channel.invokeMethod<void>('AlarmService.initialized');
 }
 
+// A lambda that returns the current instant in the form of a [DateTime].
+typedef DateTime _Now();
+// A lambda that gets the handle for the given [callback].
+typedef CallbackHandle _GetCallbackHandle(Function callback);
+
 /// A Flutter plugin for registering Dart callbacks with the Android
 /// AlarmManager service.
 ///
 /// See the example/ directory in this package for sample usage.
 class AndroidAlarmManager {
   static const String _channelName = 'plugins.flutter.io/android_alarm_manager';
-  static const MethodChannel _channel =
-      MethodChannel(_channelName, JSONMethodCodec());
+  static MethodChannel _channel =
+      const MethodChannel(_channelName, JSONMethodCodec());
+  // Function used to get the current time. It's [DateTime.now] by default.
+  static _Now _now = () => DateTime.now();
+  // Callback used to get the handle for a callback. It's
+  // [PluginUtilities.getCallbackHandle] by default.
+  static _GetCallbackHandle _getCallbackHandle =
+      (Function callback) => PluginUtilities.getCallbackHandle(callback);
+
+  /// This is exposed for the unit tests. It should not be accessed by users of
+  /// the plugin.
+  @visibleForTesting
+  static void setTestOverides(
+      {_Now now, _GetCallbackHandle getCallbackHandle}) {
+    _now = (now ?? _now);
+    _getCallbackHandle = (getCallbackHandle ?? _getCallbackHandle);
+  }
 
   /// Starts the [AndroidAlarmManager] service. This must be called before
   /// setting any alarms.
@@ -68,7 +89,7 @@ class AndroidAlarmManager {
   /// failure.
   static Future<bool> initialize() async {
     final CallbackHandle handle =
-        PluginUtilities.getCallbackHandle(_alarmManagerCallbackDispatcher);
+        _getCallbackHandle(_alarmManagerCallbackDispatcher);
     if (handle == null) {
       return false;
     }
@@ -125,7 +146,7 @@ class AndroidAlarmManager {
     bool rescheduleOnReboot = false,
   }) =>
       oneShotAt(
-        DateTime.now().add(delay),
+        _now().add(delay),
         id,
         callback,
         alarmClock: alarmClock,
@@ -182,10 +203,11 @@ class AndroidAlarmManager {
     bool wakeup = false,
     bool rescheduleOnReboot = false,
   }) async {
+    // ignore: inference_failure_on_function_return_type
     assert(callback is Function() || callback is Function(int));
     assert(id.bitLength < 32);
     final int startMillis = time.millisecondsSinceEpoch;
-    final CallbackHandle handle = PluginUtilities.getCallbackHandle(callback);
+    final CallbackHandle handle = _getCallbackHandle(callback);
     if (handle == null) {
       return false;
     }
@@ -245,13 +267,14 @@ class AndroidAlarmManager {
     bool wakeup = false,
     bool rescheduleOnReboot = false,
   }) async {
+    // ignore: inference_failure_on_function_return_type
     assert(callback is Function() || callback is Function(int));
     assert(id.bitLength < 32);
-    final int now = DateTime.now().millisecondsSinceEpoch;
+    final int now = _now().millisecondsSinceEpoch;
     final int period = duration.inMilliseconds;
     final int first =
         startAt != null ? startAt.millisecondsSinceEpoch : now + period;
-    final CallbackHandle handle = PluginUtilities.getCallbackHandle(callback);
+    final CallbackHandle handle = _getCallbackHandle(callback);
     if (handle == null) {
       return false;
     }
