@@ -15,8 +15,6 @@
 @property(nullable, copy, nonatomic) ShouldAddStorePayment shouldAddStorePayment;
 @property(nullable, copy, nonatomic) UpdatedDownloads updatedDownloads;
 
-@property(strong, nonatomic) NSMutableDictionary *transactionsSetter;
-
 @end
 
 @implementation FIAPaymentQueueHandler
@@ -31,21 +29,29 @@
                         updatedDownloads:(nullable UpdatedDownloads)updatedDownloads {
   self = [super init];
   if (self) {
-    self.queue = queue;
-    self.transactionsUpdated = transactionsUpdated;
-    self.transactionsRemoved = transactionsRemoved;
-    self.restoreTransactionFailed = restoreTransactionFailed;
-    self.paymentQueueRestoreCompletedTransactionsFinished = restoreCompletedTransactionsFinished;
-    self.shouldAddStorePayment = shouldAddStorePayment;
-    self.updatedDownloads = updatedDownloads;
-    self.transactionsSetter = [NSMutableDictionary new];
-    [queue addTransactionObserver:self];
+    _queue = queue;
+    _transactionsUpdated = transactionsUpdated;
+    _transactionsRemoved = transactionsRemoved;
+    _restoreTransactionFailed = restoreTransactionFailed;
+    _paymentQueueRestoreCompletedTransactionsFinished = restoreCompletedTransactionsFinished;
+    _shouldAddStorePayment = shouldAddStorePayment;
+    _updatedDownloads = updatedDownloads;
   }
   return self;
 }
 
-- (void)addPayment:(SKPayment *)payment {
+- (void)startObservingPaymentQueue {
+  [_queue addTransactionObserver:self];
+}
+
+- (BOOL)addPayment:(SKPayment *)payment {
+  for (SKPaymentTransaction *transaction in self.queue.transactions) {
+    if ([transaction.payment.productIdentifier isEqualToString:payment.productIdentifier]) {
+      return NO;
+    }
+  }
   [self.queue addPayment:payment];
+  return YES;
 }
 
 - (void)finishTransaction:(SKPaymentTransaction *)transaction {
@@ -61,15 +67,11 @@
 }
 
 #pragma mark - observing
+
 // Sent when the transaction array has changed (additions or state changes).  Client should check
 // state of transactions and finish as appropriate.
 - (void)paymentQueue:(SKPaymentQueue *)queue
     updatedTransactions:(NSArray<SKPaymentTransaction *> *)transactions {
-  for (SKPaymentTransaction *transaction in transactions) {
-    if (transaction.transactionIdentifier) {
-      [self.transactionsSetter setObject:transaction forKey:transaction.transactionIdentifier];
-    }
-  }
   // notify dart through callbacks.
   self.transactionsUpdated(transactions);
 }
@@ -105,10 +107,8 @@
   return (self.shouldAddStorePayment(payment, product));
 }
 
-#pragma mark - getter
-
-- (NSDictionary *)transactions {
-  return self.transactionsSetter;
+- (NSArray<SKPaymentTransaction *> *)getUnfinishedTransactions {
+  return self.queue.transactions;
 }
 
 @end
