@@ -35,10 +35,14 @@ class SKPaymentQueueWrapper {
     return _singleton;
   }
 
-  static final SKPaymentQueueWrapper _singleton = new SKPaymentQueueWrapper._();
+  static final SKPaymentQueueWrapper _singleton = SKPaymentQueueWrapper._();
 
-  SKPaymentQueueWrapper._() {
-    callbackChannel.setMethodCallHandler(_handleObserverCallbacks);
+  SKPaymentQueueWrapper._();
+
+  /// Calls [`-[SKPaymentQueue transactions]`](https://developer.apple.com/documentation/storekit/skpaymentqueue/1506026-transactions?language=objc)
+  Future<List<SKPaymentTransactionWrapper>> transactions() async {
+    return _getTransactionList(
+        await channel.invokeListMethod<Map>('-[SKPaymentQueue transactions]'));
   }
 
   /// Calls [`-[SKPaymentQueue canMakePayments:]`](https://developer.apple.com/documentation/storekit/skpaymentqueue/1506139-canmakepayments?language=objc).
@@ -53,6 +57,7 @@ class SKPaymentQueueWrapper {
   /// addTransactionObserver:]`](https://developer.apple.com/documentation/storekit/skpaymentqueue/1506042-addtransactionobserver?language=objc).
   void setTransactionObserver(SKTransactionObserverWrapper observer) {
     _observer = observer;
+    callbackChannel.setMethodCallHandler(_handleObserverCallbacks);
   }
 
   /// Posts a payment to the queue.
@@ -98,9 +103,11 @@ class SKPaymentQueueWrapper {
   /// finishTransaction:]`](https://developer.apple.com/documentation/storekit/skpaymentqueue/1506003-finishtransaction?language=objc).
   Future<void> finishTransaction(
       SKPaymentTransactionWrapper transaction) async {
+    Map<String, String> requestMap = transaction.toFinishMap();
     await channel.invokeMethod<void>(
-        '-[InAppPurchasePlugin finishTransaction:result:]',
-        transaction.transactionIdentifier);
+      '-[InAppPurchasePlugin finishTransaction:result:]',
+      requestMap,
+    );
   }
 
   /// Restore previously purchased transactions.
@@ -196,6 +203,7 @@ class SKPaymentQueueWrapper {
 /// [NSError](https://developer.apple.com/documentation/foundation/nserror?language=objc).
 @JsonSerializable(nullable: true)
 class SKError {
+  /// Creates a new [SKError] object with the provided information.
   SKError(
       {@required this.code, @required this.domain, @required this.userInfo});
 
@@ -251,6 +259,7 @@ class SKError {
 /// initiate a payment.
 @JsonSerializable(nullable: true)
 class SKPaymentWrapper {
+  /// Creates a new [SKPaymentWrapper] with the provided information.
   SKPaymentWrapper(
       {@required this.productIdentifier,
       this.applicationUsername,
