@@ -6,6 +6,7 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart' show TestWidgetsFlutterBinding;
 import 'package:in_app_purchase/src/in_app_purchase/purchase_details.dart';
 import 'package:test/test.dart';
 
@@ -17,6 +18,8 @@ import 'package:in_app_purchase/store_kit_wrappers.dart';
 import '../store_kit_wrappers/sk_test_stub_objects.dart';
 
 void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
   final FakeIOSPlatform fakeIOSPlatform = FakeIOSPlatform();
 
   setUpAll(() {
@@ -84,6 +87,28 @@ void main() {
       expect(
           response.pastPurchases.first.verificationData.serverVerificationData,
           'dummy base64data');
+      expect(response.error, isNull);
+    });
+
+    test('queryPastPurchases should not block transaction updates', () async {
+      fakeIOSPlatform.transactions
+          .add(fakeIOSPlatform.createPurchasedTransaction('foo', 'bar'));
+      Completer completer = Completer();
+      Stream<List<PurchaseDetails>> stream =
+          AppStoreConnection.instance.purchaseUpdatedStream;
+
+      StreamSubscription subscription;
+      subscription = stream.listen((purchaseDetailsList) {
+        if (purchaseDetailsList.first.status == PurchaseStatus.purchased) {
+          completer.complete(purchaseDetailsList);
+          subscription.cancel();
+        }
+      });
+      QueryPurchaseDetailsResponse response =
+          await AppStoreConnection.instance.queryPastPurchases();
+      List<PurchaseDetails> result = await completer.future;
+      expect(result.length, 1);
+      expect(result.first.productID, 'foo');
       expect(response.error, isNull);
     });
 
@@ -225,7 +250,7 @@ void main() {
           .buyNonConsumable(purchaseParam: purchaseParam);
 
       IAPError completerError = await completer.future;
-      expect(completerError.code, kPurchaseErrorCode);
+      expect(completerError.code, 'purchase_error');
       expect(completerError.source, IAPSource.AppStore);
       expect(completerError.message, 'ios_domain');
       expect(completerError.details, {'message': 'an error message'});
@@ -242,7 +267,7 @@ void main() {
       subscription = stream.listen((purchaseDetailsList) {
         details.addAll(purchaseDetailsList);
         purchaseDetailsList.forEach((purchaseDetails) {
-          if (purchaseDetails.status == PurchaseStatus.purchased) {
+          if (purchaseDetails.pendingCompletePurchase) {
             AppStoreConnection.instance.completePurchase(purchaseDetails);
             completer.complete(details);
             subscription.cancel();
@@ -323,33 +348,33 @@ class FakeIOSPlatform {
     testRestoredError = null;
   }
 
-  SKPaymentTransactionWrapper createPendingTransactionWithProductID(String id) {
+  SKPaymentTransactionWrapper createPendingTransaction(String id) {
     return SKPaymentTransactionWrapper(
+        transactionIdentifier: null,
         payment: SKPaymentWrapper(productIdentifier: id),
         transactionState: SKPaymentTransactionStateWrapper.purchasing,
         transactionTimeStamp: 123123.121,
-        transactionIdentifier: id,
         error: null,
         originalTransaction: null);
   }
 
-  SKPaymentTransactionWrapper createPurchasedTransactionWithProductID(
-      String id) {
+  SKPaymentTransactionWrapper createPurchasedTransaction(
+      String productId, String transactionId) {
     return SKPaymentTransactionWrapper(
-        payment: SKPaymentWrapper(productIdentifier: id),
+        payment: SKPaymentWrapper(productIdentifier: productId),
         transactionState: SKPaymentTransactionStateWrapper.purchased,
         transactionTimeStamp: 123123.121,
-        transactionIdentifier: id,
+        transactionIdentifier: transactionId,
         error: null,
         originalTransaction: null);
   }
 
-  SKPaymentTransactionWrapper createFailedTransactionWithProductID(String id) {
+  SKPaymentTransactionWrapper createFailedTransaction(String productId) {
     return SKPaymentTransactionWrapper(
-        payment: SKPaymentWrapper(productIdentifier: id),
+        transactionIdentifier: null,
+        payment: SKPaymentWrapper(productIdentifier: productId),
         transactionState: SKPaymentTransactionStateWrapper.failed,
         transactionTimeStamp: 123123.121,
-        transactionIdentifier: id,
         error: SKError(
             code: 0,
             domain: 'ios_domain',
@@ -409,26 +434,26 @@ class FakeIOSPlatform {
         return Future<void>.sync(() {});
       case '-[InAppPurchasePlugin addPayment:result:]':
         String id = call.arguments['productIdentifier'];
-        SKPaymentTransactionWrapper transaction =
-            createPendingTransactionWithProductID(id);
+        SKPaymentTransactionWrapper transaction = createPendingTransaction(id);
         AppStoreConnection.observer
             .updatedTransactions(transactions: [transaction]);
         sleep(const Duration(milliseconds: 30));
         if (testTransactionFail) {
           SKPaymentTransactionWrapper transaction_failed =
-              createFailedTransactionWithProductID(id);
+              createFailedTransaction(id);
           AppStoreConnection.observer
               .updatedTransactions(transactions: [transaction_failed]);
         } else {
           SKPaymentTransactionWrapper transaction_finished =
-              createPurchasedTransactionWithProductID(id);
+              createPurchasedTransaction(id, transaction.transactionIdentifier);
           AppStoreConnection.observer
               .updatedTransactions(transactions: [transaction_finished]);
         }
         break;
       case '-[InAppPurchasePlugin finishTransaction:result:]':
-        finishedTransactions
-            .add(createPurchasedTransactionWithProductID(call.arguments));
+        finishedTransactions.add(createPurchasedTransaction(
+            call.arguments["productIdentifier"],
+            call.arguments["transactionIdentifier"]));
         break;
     }
     return Future<void>.sync(() {});
