@@ -2,8 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// @dart = 2.9
+
+import 'dart:io';
 import 'dart:ui';
 
+import 'package:flutter_test/flutter_test.dart' show TestWidgetsFlutterBinding;
 import 'package:mockito/mockito.dart';
 import 'package:share/share.dart';
 import 'package:test/test.dart';
@@ -11,6 +15,8 @@ import 'package:test/test.dart';
 import 'package:flutter/services.dart';
 
 void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
   MockMethodChannel mockChannel;
 
   setUp(() {
@@ -18,16 +24,8 @@ void main() {
     // Re-pipe to mockito for easier verifies.
     Share.channel.setMockMethodCallHandler((MethodCall call) async {
       // The explicit type can be void as the only method call has a return type of void.
-      mockChannel.invokeMethod<void>(call.method, call.arguments);
+      await mockChannel.invokeMethod<void>(call.method, call.arguments);
     });
-  });
-
-  test('sharing null fails', () {
-    expect(
-      () => Share.share(null),
-      throwsA(const TypeMatcher<AssertionError>()),
-    );
-    verifyZeroInteractions(mockChannel);
   });
 
   test('sharing empty fails', () {
@@ -52,6 +50,44 @@ void main() {
       'originWidth': 3.0,
       'originHeight': 4.0,
     }));
+  });
+
+  test('sharing empty file fails', () {
+    expect(
+      () => Share.shareFiles(['']),
+      throwsA(const TypeMatcher<AssertionError>()),
+    );
+    verifyZeroInteractions(mockChannel);
+  });
+
+  test('sharing file sets correct mimeType', () async {
+    final String path = 'tempfile-83649a.png';
+    final File file = File(path);
+    try {
+      file.createSync();
+      await Share.shareFiles([path]);
+      verify(mockChannel.invokeMethod('shareFiles', <String, dynamic>{
+        'paths': [path],
+        'mimeTypes': ['image/png'],
+      }));
+    } finally {
+      file.deleteSync();
+    }
+  });
+
+  test('sharing file sets passed mimeType', () async {
+    final String path = 'tempfile-83649a.png';
+    final File file = File(path);
+    try {
+      file.createSync();
+      await Share.shareFiles([path], mimeTypes: ['*/*']);
+      verify(mockChannel.invokeMethod('shareFiles', <String, dynamic>{
+        'paths': [file.path],
+        'mimeTypes': ['*/*'],
+      }));
+    } finally {
+      file.deleteSync();
+    }
   });
 }
 
