@@ -5,10 +5,10 @@
 #import "FLTWKNavigationDelegate.h"
 
 @implementation FLTWKNavigationDelegate {
-  FlutterMethodChannel* _methodChannel;
+  FlutterMethodChannel *_methodChannel;
 }
 
-- (instancetype)initWithChannel:(FlutterMethodChannel*)channel {
+- (instancetype)initWithChannel:(FlutterMethodChannel *)channel {
   self = [super init];
   if (self) {
     _methodChannel = channel;
@@ -16,14 +16,20 @@
   return self;
 }
 
-- (void)webView:(WKWebView*)webView
-    decidePolicyForNavigationAction:(WKNavigationAction*)navigationAction
+#pragma mark - WKNavigationDelegate conformance
+
+- (void)webView:(WKWebView *)webView didStartProvisionalNavigation:(WKNavigation *)navigation {
+  [_methodChannel invokeMethod:@"onPageStarted" arguments:@{@"url" : webView.URL.absoluteString}];
+}
+
+- (void)webView:(WKWebView *)webView
+    decidePolicyForNavigationAction:(WKNavigationAction *)navigationAction
                     decisionHandler:(void (^)(WKNavigationActionPolicy))decisionHandler {
   if (!self.hasDartNavigationDelegate) {
     decisionHandler(WKNavigationActionPolicyAllow);
     return;
   }
-  NSDictionary* arguments = @{
+  NSDictionary *arguments = @{
     @"url" : navigationAction.request.URL.absoluteString,
     @"isForMainFrame" : @(navigationAction.targetFrame.isMainFrame)
   };
@@ -50,13 +56,61 @@
                             decisionHandler(WKNavigationActionPolicyAllow);
                             return;
                           }
-                          NSNumber* typedResult = result;
+                          NSNumber *typedResult = result;
                           decisionHandler([typedResult boolValue] ? WKNavigationActionPolicyAllow
                                                                   : WKNavigationActionPolicyCancel);
                         }];
 }
 
-- (void)webView:(WKWebView*)webView didFinishNavigation:(WKNavigation*)navigation {
+- (void)webView:(WKWebView *)webView didFinishNavigation:(WKNavigation *)navigation {
   [_methodChannel invokeMethod:@"onPageFinished" arguments:@{@"url" : webView.URL.absoluteString}];
 }
+
++ (id)errorCodeToString:(NSUInteger)code {
+  switch (code) {
+    case WKErrorUnknown:
+      return @"unknown";
+    case WKErrorWebContentProcessTerminated:
+      return @"webContentProcessTerminated";
+    case WKErrorWebViewInvalidated:
+      return @"webViewInvalidated";
+    case WKErrorJavaScriptExceptionOccurred:
+      return @"javaScriptExceptionOccurred";
+    case WKErrorJavaScriptResultTypeIsUnsupported:
+      return @"javaScriptResultTypeIsUnsupported";
+  }
+
+  return [NSNull null];
+}
+
+- (void)onWebResourceError:(NSError *)error {
+  [_methodChannel invokeMethod:@"onWebResourceError"
+                     arguments:@{
+                       @"errorCode" : @(error.code),
+                       @"domain" : error.domain,
+                       @"description" : error.description,
+                       @"errorType" : [FLTWKNavigationDelegate errorCodeToString:error.code],
+                     }];
+}
+
+- (void)webView:(WKWebView *)webView
+    didFailNavigation:(WKNavigation *)navigation
+            withError:(NSError *)error {
+  [self onWebResourceError:error];
+}
+
+- (void)webView:(WKWebView *)webView
+    didFailProvisionalNavigation:(WKNavigation *)navigation
+                       withError:(NSError *)error {
+  [self onWebResourceError:error];
+}
+
+- (void)webViewWebContentProcessDidTerminate:(WKWebView *)webView {
+  NSError *contentProcessTerminatedError =
+      [[NSError alloc] initWithDomain:WKErrorDomain
+                                 code:WKErrorWebContentProcessTerminated
+                             userInfo:nil];
+  [self onWebResourceError:contentProcessTerminatedError];
+}
+
 @end
