@@ -10,6 +10,8 @@ import '../platform_interface.dart';
 
 /// A [WebViewPlatformController] that uses a method channel to control the webview.
 class MethodChannelWebViewPlatform implements WebViewPlatformController {
+  /// Constructs an instance that will listen for webviews broadcasting to the
+  /// given [id], using the given [WebViewPlatformCallbacksHandler].
   MethodChannelWebViewPlatform(int id, this._platformCallbacksHandler)
       : assert(_platformCallbacksHandler != null),
         _channel = MethodChannel('plugins.flutter.io/webview_$id') {
@@ -23,30 +25,54 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
   static const MethodChannel _cookieManagerChannel =
       MethodChannel('plugins.flutter.io/cookie_manager');
 
-  Future<bool> _onMethodCall(MethodCall call) async {
+  Future<bool?> _onMethodCall(MethodCall call) async {
     switch (call.method) {
       case 'javascriptChannelMessage':
-        final String channel = call.arguments['channel'];
-        final String message = call.arguments['message'];
+        final String channel = call.arguments['channel']!;
+        final String message = call.arguments['message']!;
         _platformCallbacksHandler.onJavaScriptChannelMessage(channel, message);
         return true;
       case 'navigationRequest':
-        return _platformCallbacksHandler.onNavigationRequest(
-          url: call.arguments['url'],
-          isForMainFrame: call.arguments['isForMainFrame'],
+        return await _platformCallbacksHandler.onNavigationRequest(
+          url: call.arguments['url']!,
+          isForMainFrame: call.arguments['isForMainFrame']!,
         );
       case 'onPageFinished':
-        _platformCallbacksHandler.onPageFinished(call.arguments['url']);
+        _platformCallbacksHandler.onPageFinished(call.arguments['url']!);
+        return null;
+      case 'onPageStarted':
+        _platformCallbacksHandler.onPageStarted(call.arguments['url']!);
+        return null;
+      case 'onWebResourceError':
+        _platformCallbacksHandler.onWebResourceError(
+          WebResourceError(
+            errorCode: call.arguments['errorCode']!,
+            description: call.arguments['description']!,
+            // iOS doesn't support `failingUrl`.
+            failingUrl: call.arguments['failingUrl'],
+            domain: call.arguments['domain'],
+            errorType: call.arguments['errorType'] == null
+                ? null
+                : WebResourceErrorType.values.firstWhere(
+                    (WebResourceErrorType type) {
+                      return type.toString() ==
+                          '$WebResourceErrorType.${call.arguments['errorType']}';
+                    },
+                  ),
+          ),
+        );
         return null;
     }
+
     throw MissingPluginException(
-        '${call.method} was invoked but has no handler');
+      '${call.method} was invoked but has no handler',
+    );
   }
 
   @override
   Future<void> loadUrl(
     String url,
-    Map<String, String> headers,
+    Map<String, String>? headers,
   ) async {
     assert(url != null);
     return _channel.invokeMethod<void>('loadUrl', <String, dynamic>{
@@ -56,13 +82,15 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
   }
 
   @override
-  Future<String> currentUrl() => _channel.invokeMethod<String>('currentUrl');
+  Future<String?> currentUrl() => _channel.invokeMethod<String>('currentUrl');
 
   @override
-  Future<bool> canGoBack() => _channel.invokeMethod<bool>("canGoBack");
+  Future<bool> canGoBack() =>
+      _channel.invokeMethod<bool>("canGoBack").then((result) => result!);
 
   @override
-  Future<bool> canGoForward() => _channel.invokeMethod<bool>("canGoForward");
+  Future<bool> canGoForward() =>
+      _channel.invokeMethod<bool>("canGoForward").then((result) => result!);
 
   @override
   Future<void> goBack() => _channel.invokeMethod<void>("goBack");
@@ -77,18 +105,18 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
   Future<void> clearCache() => _channel.invokeMethod<void>("clearCache");
 
   @override
-  Future<void> updateSettings(WebSettings settings) {
+  Future<void> updateSettings(WebSettings settings) async {
     final Map<String, dynamic> updatesMap = _webSettingsToMap(settings);
-    if (updatesMap.isEmpty) {
-      return null;
+    if (updatesMap.isNotEmpty) {
+      await _channel.invokeMethod<void>('updateSettings', updatesMap);
     }
-    return _channel.invokeMethod<void>('updateSettings', updatesMap);
   }
 
   @override
   Future<String> evaluateJavascript(String javascriptString) {
-    return _channel.invokeMethod<String>(
-        'evaluateJavascript', javascriptString);
+    return _channel
+        .invokeMethod<String>('evaluateJavascript', javascriptString)
+        .then((result) => result!);
   }
 
   @override
@@ -103,14 +131,41 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
         'removeJavascriptChannels', javascriptChannelNames.toList());
   }
 
+  @override
+  Future<String?> getTitle() => _channel.invokeMethod<String>("getTitle");
+
+  @override
+  Future<void> scrollTo(int x, int y) {
+    return _channel.invokeMethod<void>('scrollTo', <String, int>{
+      'x': x,
+      'y': y,
+    });
+  }
+
+  @override
+  Future<void> scrollBy(int x, int y) {
+    return _channel.invokeMethod<void>('scrollBy', <String, int>{
+      'x': x,
+      'y': y,
+    });
+  }
+
+  @override
+  Future<int> getScrollX() =>
+      _channel.invokeMethod<int>("getScrollX").then((result) => result!);
+
+  @override
+  Future<int> getScrollY() =>
+      _channel.invokeMethod<int>("getScrollY").then((result) => result!);
+
   /// Method channel implementation for [WebViewPlatform.clearCookies].
   static Future<bool> clearCookies() {
     return _cookieManagerChannel
         .invokeMethod<bool>('clearCookies')
-        .then<bool>((dynamic result) => result);
+        .then<bool>((dynamic result) => result!);
   }
 
-  static Map<String, dynamic> _webSettingsToMap(WebSettings settings) {
+  static Map<String, dynamic> _webSettingsToMap(WebSettings? settings) {
     final Map<String, dynamic> map = <String, dynamic>{};
     void _addIfNonNull(String key, dynamic value) {
       if (value == null) {
@@ -126,9 +181,13 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
       map[key] = setting.value;
     }
 
-    _addIfNonNull('jsMode', settings.javascriptMode?.index);
+    _addIfNonNull('jsMode', settings!.javascriptMode?.index);
     _addIfNonNull('hasNavigationDelegate', settings.hasNavigationDelegate);
     _addIfNonNull('debuggingEnabled', settings.debuggingEnabled);
+    _addIfNonNull(
+        'gestureNavigationEnabled', settings.gestureNavigationEnabled);
+    _addIfNonNull(
+        'allowsInlineMediaPlayback', settings.allowsInlineMediaPlayback);
     _addSettingIfPresent('userAgent', settings.userAgent);
     return map;
   }
